+++ orchestration/src/main/java/com/sap/ai/sdk/orchestration/OrchestrationAiModel.java
@@ -86,6 +86,10 @@ public class OrchestrationAiModel {
   public static final OrchestrationAiModel MIXTRAL_8X7B_INSTRUCT_V01 =
       new OrchestrationAiModel("mistralai--mixtral-8x7b-instruct-v01");
 
+  /** MistralAI Mistral Small model */
+  public static final OrchestrationAiModel MISTRAL_SMALL =
+      new OrchestrationAiModel("mistralai--mistral-small");
+
   /**
    * Meta Llama3 70B Instruct model
    *
@@ -178,6 +182,10 @@ public class OrchestrationAiModel {
   /** Anthropic Claude 4.6 Sonnet model */
   public static final OrchestrationAiModel CLAUDE_4_6_SONNET =
       new OrchestrationAiModel("anthropic--claude-4.6-sonnet");
+
+  /** Anthropic Claude 4.7 Opus model */
+  public static final OrchestrationAiModel CLAUDE_4_7_OPUS =
+      new OrchestrationAiModel("anthropic--claude-4.7-opus");
 
   /**
    * Amazon Titan Text Lite model
@@ -393,6 +401,10 @@ public class OrchestrationAiModel {
   /** Google Cloud Platform Gemini 2.5 Pro model */
   public static final OrchestrationAiModel GEMINI_2_5_PRO =
       new OrchestrationAiModel("gemini-2.5-pro");
+
+  /** Google Cloud Platform Gemini 3.1 Flash Lite model */
+  public static final OrchestrationAiModel GEMINI_3_1_FLASH_LITE =
+      new OrchestrationAiModel("gemini-3.1-flash-lite");
 
   /**
    * Alephalpha-pharia-1-7b-control model