+++ sample-code/spring-app/src/main/java/com/sap/ai/sdk/app/controllers/PromptRegistryController.java
@@ -32,6 +32,7 @@ import java.util.Map;
 import lombok.val;
 import org.springframework.ai.chat.client.ChatClient;
 import org.springframework.ai.chat.client.advisor.MessageChatMemoryAdvisor;
+import org.springframework.ai.chat.memory.ChatMemory;
 import org.springframework.ai.chat.memory.InMemoryChatMemoryRepository;
 import org.springframework.ai.chat.memory.MessageWindowChatMemory;
 import org.springframework.ai.chat.messages.Message;
@@ -93,7 +94,7 @@ class PromptRegistryController {
 
   @GetMapping("/importTemplate")
   PromptTemplatePostResponse importTemplate() throws IOException {
-    val template = new ClassPathResource("prompt-template.yaml").getContentAsByteArray();
+    val template = new ClassPathResource("prompt-template.yaml").getFile();
     return promptClient.importPromptTemplate("default", null, template);
   }
 
@@ -147,7 +148,11 @@ class PromptRegistryController {
 
     final List<Message> messages = SpringAiConverter.promptTemplateToMessages(promptResponse);
     val prompt = new Prompt(messages);
-    val response = cl.prompt(prompt).call().chatResponse();
+    val response =
+        cl.prompt(prompt)
+            .advisors(spec -> spec.param(ChatMemory.CONVERSATION_ID, "conversation id"))
+            .call()
+            .chatResponse();
     return response != null ? response.getResult() : null;
   }